Major refactoring of the GameActivity.
