+++ app/src/main/java/com/github/polypoly/app/base/game/Game.kt
@@ -25,9 +25,9 @@ class Game private constructor(
     val dateBegin: Long = System.currentTimeMillis(),
 ) {
 
-    private val inGameLocations: List<InGameLocation> = rules.gameMap.flatMap { zone -> zone.locations.map { location ->
+    private val inGameLocations: List<InGameLocation> = rules.gameMap.flatMap { zone -> zone.locationProperties.map { location ->
         InGameLocation(
-            location = location,
+            locationProperty = location,
             owner = null,
             level = LocalizationLevel.LEVEL_0,
         ) } }
