+++ app/src/main/java/com/github/polypoly/app/base/game/location/InGameLocation.kt
@@ -4,14 +4,14 @@ import com.github.polypoly.app.base.game.Player
 
 /**
  * A class that represent an location in a game
- * @property location The location of the in-game location, that permit to know the static info of
+ * @property locationProperty The location of the in-game location, that permit to know the static info of
  * the location as the name, the base price, the base tax price, etc...
  * @property level The current level of the location that determine the price and the tax price
  * @property owner The owner of the location, if the location is not owned, the owner is null
  * @property bids The list of the bids made on the location by the players in the previous round
  */
 data class InGameLocation (
-    var location: Location,
+    var locationProperty: LocationProperty,
     var level: LocalizationLevel = LocalizationLevel.LEVEL_0,
     var owner: Player? = null,
 ) {
@@ -32,7 +32,7 @@ data class InGameLocation (
      * @return the current tax of the location
      */
     fun currentTax(): Int {
-        return location.baseTaxPrice * (level.ordinal+1)
+        return locationProperty.baseTaxPrice * (level.ordinal+1)
     }
 
     /**
@@ -40,7 +40,7 @@ data class InGameLocation (
      * @return the current price of the location
      */
     fun currentPrice(): Int {
-        return location.basePrice * (level.ordinal+1)
+        return locationProperty.basePrice * (level.ordinal+1)
     }
 
     /**
@@ -48,7 +48,7 @@ data class InGameLocation (
      * @return the current mortgage price of the location
      */
     fun currentMortgagePrice(): Int {
-        return location.baseMortgagePrice * (level.ordinal + 1)
+        return locationProperty.baseMortgagePrice * (level.ordinal + 1)
     }
 
     /**
