+++ app/src/main/java/com/github/polypoly/app/base/game/location/LocationRepository.kt
@@ -39,8 +39,8 @@ object LocationRepository {
     /**
      * List of localizations in the EPFL campus area.
      */
-    private val locations = listOf(
-        Location(
+    private val locationProperties = listOf(
+        LocationProperty(
             name = "BC",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -48,7 +48,7 @@ object LocationRepository {
             latitude = 46.51861304708622,
             longitude = 6.561904544895297,
         ),
-        Location(
+        LocationProperty(
             name = "INM",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -56,7 +56,7 @@ object LocationRepository {
             latitude = 46.51862967945017,
             longitude = 6.563195429654059,
         ),
-        Location(
+        LocationProperty(
             name = "INF",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -64,7 +64,7 @@ object LocationRepository {
             latitude = 46.5187569432044,
             longitude = 6.563754302940894,
         ),
-        Location(
+        LocationProperty(
             name = "INJ",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -72,7 +72,7 @@ object LocationRepository {
             latitude = 46.5183787509451,
             longitude = 6.563782977860895,
         ),
-        Location(
+        LocationProperty(
             name = "INN",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -80,7 +80,7 @@ object LocationRepository {
             latitude = 46.5187602318236,
             longitude = 6.562524865762242,
         ),
-        Location(
+        LocationProperty(
             name = "INR",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -88,7 +88,7 @@ object LocationRepository {
             latitude = 46.51837464014728,
             longitude = 6.562565488564992,
         ),
-        Location(
+        LocationProperty(
             name = "QIG",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -96,7 +96,7 @@ object LocationRepository {
             latitude = 46.51757926537674,
             longitude = 6.561402515609107,
         ),
-        Location(
+        LocationProperty(
             name = "QIH",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -104,7 +104,7 @@ object LocationRepository {
             latitude= 46.51682504051627,
             longitude = 6.5608779628377905,
         ),
-        Location(
+        LocationProperty(
             name = "QII",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -112,7 +112,7 @@ object LocationRepository {
             latitude = 46.516334883735226,
             longitude = 6.561316930683261,
         ),
-        Location(
+        LocationProperty(
             name = "QIJ",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -120,7 +120,7 @@ object LocationRepository {
             latitude = 46.51641847667725,
             longitude = 6.562045782954984,
         ),
-        Location(
+        LocationProperty(
             name = "QIF",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -128,7 +128,7 @@ object LocationRepository {
             latitude = 46.51678956668743,
             longitude = 6.56166017548698,
         ),
-        Location(
+        LocationProperty(
             name = "Logitech",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -136,7 +136,7 @@ object LocationRepository {
             latitude = 46.5165653868565,
             longitude = 6.562648543340301,
         ),
-        Location(
+        LocationProperty(
             name = "PSE D",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -144,7 +144,7 @@ object LocationRepository {
             latitude = 46.51735208130024,
             longitude = 6.561982124516758,
         ),
-        Location(
+        LocationProperty(
             name = "PSE C",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -152,7 +152,7 @@ object LocationRepository {
             latitude = 46.51739626509118,
             longitude = 6.562762759441413,
         ),
-        Location(
+        LocationProperty(
             name = "PSE A",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -160,7 +160,7 @@ object LocationRepository {
             latitude = 46.51777097627283,
             longitude = 6.5625767820914955,
         ),
-        Location(
+        LocationProperty(
             name = "PPH",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -168,7 +168,7 @@ object LocationRepository {
             latitude = 46.51768958561605,
             longitude = 6.564391504839671,
         ),
-        Location(
+        LocationProperty(
             name = "PPB",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -176,7 +176,7 @@ object LocationRepository {
             latitude = 46.517154729671965,
             longitude = 6.564712545176651,
         ),
-        Location(
+        LocationProperty(
             name = "TCV",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -184,7 +184,7 @@ object LocationRepository {
             latitude = 46.51764075116125,
             longitude = 6.565030206141661,
         ),
-        Location(
+        LocationProperty(
             name = "Rolex Learning Center",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -192,7 +192,7 @@ object LocationRepository {
             latitude = 46.518370937968015,
             longitude = 6.568348749416791,
         ),
-        Location(
+        LocationProperty(
             name = "ELG",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -200,7 +200,7 @@ object LocationRepository {
             latitude = 46.51877090875509,
             longitude = 6.564442195419994,
         ),
-        Location(
+        LocationProperty(
             name = "ELH", // NORTH WEST BUILDING (FAKE TEMPORARY NAME)
             basePrice = 200,
             baseTaxPrice = 20,
@@ -208,7 +208,7 @@ object LocationRepository {
             latitude = 46.519142491902066,
             longitude = 6.564474343002977,
         ),
-        Location(
+        LocationProperty(
             name = "ELI", // SOUTH EAST BUILDING (FAKE TEMPORARY NAME)
             basePrice = 200,
             baseTaxPrice = 20,
@@ -216,7 +216,7 @@ object LocationRepository {
             latitude = 46.518740557775104,
             longitude = 6.565082406173954,
         ),
-        Location(
+        LocationProperty(
             name = "ELE",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -224,7 +224,7 @@ object LocationRepository {
             latitude = 46.51914375981869,
             longitude = 6.56506213740159,
         ),
-        Location(
+        LocationProperty(
             name = "ELL",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -232,7 +232,7 @@ object LocationRepository {
             latitude = 46.51836017574591,
             longitude = 6.565227972811845,
         ),
-        Location(
+        LocationProperty(
             name = "ELA",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -240,7 +240,7 @@ object LocationRepository {
             latitude = 46.51948102716419,
             longitude = 6.5645167231633765,
         ),
-        Location(
+        LocationProperty(
             name = "ELB",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -248,7 +248,7 @@ object LocationRepository {
             latitude = 46.51953808346133,
             longitude = 6.565047396476193,
         ),
-        Location(
+        LocationProperty(
             name = "MX",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -256,7 +256,7 @@ object LocationRepository {
             latitude = 46.5194856751895,
             longitude = 6.5631361906295265,
         ),
-        Location(
+        LocationProperty(
             name = "SV",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -264,7 +264,7 @@ object LocationRepository {
             latitude = 46.52007453561807,
             longitude = 6.563888081639078,
         ),
-        Location(
+        LocationProperty(
             name = "CO2",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -272,7 +272,7 @@ object LocationRepository {
             latitude = 46.52006946399603,
             longitude = 6.56463065575446,
         ),
-        Location(
+        LocationProperty(
             name = "CO1",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -280,7 +280,7 @@ object LocationRepository {
             latitude = 46.519989585944444,
             longitude = 6.56529768262682,
         ),
-        Location(
+        LocationProperty(
             name = "ME",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -288,7 +288,7 @@ object LocationRepository {
             latitude = 46.51957371101376,
             longitude = 6.566930240111198,
         ),
-        Location(
+        LocationProperty(
             name = "MA",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -296,7 +296,7 @@ object LocationRepository {
             latitude = 46.51997056734838,
             longitude = 6.567847862715424,
         ),
-        Location(
+        LocationProperty(
             name = "CM",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -304,7 +304,7 @@ object LocationRepository {
             latitude = 46.52049801409303,
             longitude = 6.567322717249148,
         ),
-        Location(
+        LocationProperty(
             name = "SG",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -312,7 +312,7 @@ object LocationRepository {
             latitude = 46.52093959210354,
             longitude = 6.564199076849092,
         ),
-        Location(
+        LocationProperty(
             name = "BM",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -320,7 +320,7 @@ object LocationRepository {
             latitude = 46.521132310189024,
             longitude = 6.565170135307404,
         ),
-        Location(
+        LocationProperty(
             name = "BP",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -328,7 +328,7 @@ object LocationRepository {
             latitude = 46.52145181497848,
             longitude = 6.5648052974047575,
         ),
-        Location(
+        LocationProperty(
             name = "GC",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -336,7 +336,7 @@ object LocationRepository {
             latitude = 46.52147463873658,
             longitude = 6.566549644954401,
         ),
-        Location(
+        LocationProperty(
             name = "GR",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -344,7 +344,7 @@ object LocationRepository {
             latitude = 46.52125783221227,
             longitude = 6.567854216848852,
         ),
-        Location(
+        LocationProperty(
             name = "CE",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -352,7 +352,7 @@ object LocationRepository {
             latitude = 46.520504533910575,
             longitude = 6.570065102079756,
         ),
-        Location(
+        LocationProperty(
             name = "PH",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -360,7 +360,7 @@ object LocationRepository {
             latitude = 46.521198067781015,
             longitude = 6.569884525744014,
         ),
-        Location(
+        LocationProperty(
             name = "CH",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -368,7 +368,7 @@ object LocationRepository {
             latitude = 46.51976884199191,
             longitude = 6.570168811306288,
         ),
-        Location(
+        LocationProperty(
             name = "BS",
             basePrice = 200,
             baseTaxPrice = 20,
@@ -383,108 +383,108 @@ object LocationRepository {
      */
     private val zones = listOf(
         Zone( // BC, INN, INR
-            locations = listOf(
-                locations[0],
-                locations[4],
-                locations[5]
+            locationProperties = listOf(
+                locationProperties[0],
+                locationProperties[4],
+                locationProperties[5]
             ),
             color = colours[0]
         ),
         Zone( // INM, INJ, INF
-            locations = listOf(
-                locations[3],
-                locations[2],
-                locations[1]
+            locationProperties = listOf(
+                locationProperties[3],
+                locationProperties[2],
+                locationProperties[1]
             ),
             color = colours[1]
         ),
         Zone( // QIG, QIH, QII
-            locations = listOf(
-                locations[6],
-                locations[7],
-                locations[8]
+            locationProperties = listOf(
+                locationProperties[6],
+                locationProperties[7],
+                locationProperties[8]
             ),
             color = colours[2]
         ),
         Zone( // locations with names QIF, QIJ, Logitech
-            locations = listOf(
-                locations[9],
-                locations[10],
-                locations[11]
+            locationProperties = listOf(
+                locationProperties[9],
+                locationProperties[10],
+                locationProperties[11]
             ),
             color = colours[3]
         ),
         Zone( // PSE A, PSE D, PSE C
-            locations = listOf(
-                locations[12],
-                locations[14],
-                locations[13]
+            locationProperties = listOf(
+                locationProperties[12],
+                locationProperties[14],
+                locationProperties[13]
             ),
             color = colours[4]
         ),
         Zone( // TCV, PPH, PPB
-            locations = listOf(
-                locations[15],
-                locations[17],
-                locations[16]
+            locationProperties = listOf(
+                locationProperties[15],
+                locationProperties[17],
+                locationProperties[16]
             ),
             color = colours[5]
         ),
         Zone( // ELA, ELB, ELE
-            locations = listOf(
-                locations[24],
-                locations[25],
-                locations[22]
+            locationProperties = listOf(
+                locationProperties[24],
+                locationProperties[25],
+                locationProperties[22]
             ),
             color = colours[6]
         ),
         Zone( // ELG, ELH, ELI, ELL
-            locations = listOf(
-                locations[19],
-                locations[20],
-                locations[21],
-                locations[23]
+            locationProperties = listOf(
+                locationProperties[19],
+                locationProperties[20],
+                locationProperties[21],
+                locationProperties[23]
             ),
             color = colours[7]
         ),
         Zone( // MX, SV
-            locations = listOf(
-                locations[26],
-                locations[27]
+            locationProperties = listOf(
+                locationProperties[26],
+                locationProperties[27]
             ),
             color = colours[8]
         ),
         Zone( // CO1, CO2, CM
-            locations = listOf(
-                locations[28],
-                locations[29],
-                locations[32]
+            locationProperties = listOf(
+                locationProperties[28],
+                locationProperties[29],
+                locationProperties[32]
             ),
             color = colours[9]
         ),
         Zone( // ME, MA,
-            locations = listOf(
-                locations[30],
-                locations[31]
+            locationProperties = listOf(
+                locationProperties[30],
+                locationProperties[31]
             ),
             color = colours[10]
         ),
         Zone( // SG, BM, BP, GC, GR
-            locations = listOf(
-                locations[33],
-                locations[34],
-                locations[35],
-                locations[36],
-                locations[37]
+            locationProperties = listOf(
+                locationProperties[33],
+                locationProperties[34],
+                locationProperties[35],
+                locationProperties[36],
+                locationProperties[37]
             ),
             color = colours[11]
         ),
         Zone( // CE, PH, CH, BS
-            locations = listOf(
-                locations[38],
-                locations[39],
-                locations[40],
-                locations[41]
+            locationProperties = listOf(
+                locationProperties[38],
+                locationProperties[39],
+                locationProperties[40],
+                locationProperties[41]
             ),
             color = colours[12]
         )
